Cliff-walls ChatGPT: In order to be able to test Climbers, we need cliffs
>>> Prompt: tweak of existing
...
go upwards toward a peak. Each cliff wall must be between 40 and 100 pixels wide. The bottom
of the cliff can be flat as it is always resting on a platform. The polygon you use to draw
the cliff must be filled (not just an outline).

Note that we still want to keep the existing rock and dirt platforms, they are good as they are,
don't change them. Just add the cliff walls on top of them.

You might be best off splitting the generation of platforms and cliff-walls into two separate
sections as you keep on messing up your line-drawing.

Just give me the code, no need to explain anything, I just want to paste in the whole thing
with the cliff-wall solution in it. The cliffs must be separated by at least 100 pixels on the X axis.
...

>>> Catch-up notes:
Backup was named: 'index - Copy (29) - more cliff fail.html'
Associated LLM context: n/a