A bit of cliff-walls success ChatGPT: In order to be able to test Climbers, we need cliffs
>>> Prompt: tweak of existing
...

Update the map generation below to not only give me platforms, but add steep cliff
walls here and there, these walls should only be placed on top of platforms and should have
a bit of randomly generated jagged edges. The cliff walls can be quite tall and take up quite
a bit of the full map (make this configurable).
The cliff walls should be blue (0000ff), for now.

You can picture the cliff walls like this:

    /\
   |  \
   /   \

or

/\

or

  /\
 /  \
/    \

etc.
They should look a bit jagged and randomly placed but the base of the walls must always be on top of a platform.

The code for generating the platforms is below, finish the drawCliffs method to add these cliff walls.
...

>>> Catch-up notes:
Backup was named: 'index - Copy (29) - more cliff fail.html'
Associated LLM context: n/a