Cliff-wall tweaks ChatGPT: In order to be able to test Climbers, we need cliffs
>>> Prompt: tweak of existing
...
a bit of randomly generated jagged edges. The cliff walls can be quite tall (a few platforms high at most)
and take up quite a bit of width (make this configurable). The cliff walls can be symmetrical, but they
don't have to be. The cliff walls should be blue (0000ff), for now.
...

or
 |\
 | \
 |  \
 /   \

etc.

They should look a bit jagged and randomly placed but the base of the walls must always be on top of a platform.

The code for generating the platforms is below, finish the drawCliffs method to add these cliff walls.

Just give me the new code, no need to explain anything, I just want to paste in the whole thing
with the cliff-wall solution in it.

...

>>> Catch-up notes:
Backup was named: 'index - Copy (31) - more cliff success.html'
Associated LLM context: n/a