Cliff tweaking, forbid cliffs on top ChatGPT: Prompt tweaking
I have to say I like the current cliffs (commit before this), they look really jagged and stuff. But they don't make for good testing because there are tiny holes and stuff.

>>> Prompt:
...
   ./\   /\
   |  \./  \
   /   \    \
   """"""""""
...
They should look a bit jagged and randomly placed but the base of the walls must always be on top of a platform,
however, no cliffs must be drawn on the top 50 pixels of the canvas.

>>> Catch-up notes:
Backup was named: 'index - Copy (35).html'
Associated LLM context: n/a