Bug fixing ChatGPT: Had to intervene and do manual bug-fixing
There was frustration. Once I figured out what it was trying to do, I realized it was _so_ close,
so ... I had to manually intervene.

>>> Prompt additions around climber was:

...
If a climber is climbing, you should probably not do the heightAdjustment as that only applies if the lemming
is walking on the ground.
...

...
Don't use pixelIsColor to check if pixels are climbable surfaces. Use isColorOneOf(needle, haystack) instead, where
needle is the pixel's color, and haystack is an array of several colors (e.g. the ones in terrainColorBytes).
...

>>> Catch-up notes:
Backup was named: 'index - Copy (42) - lemming 1 climbs -- but need to fix some bugs manually.html'
Associated LLM context: n/a