Minor clean-up in Lemming.update ChatGPT: see prompt
>>> Prompt:
Refactor the update function to instead of using multiple pixelIsColor(), use already implemented
isColorOneOf(needle, haystack), needle is the color we are searching for, and haystack can take
multiple colors in an array (e.g. pass in both dirt and ground in one function call).
To get a pixel's color, use the already implemented function: getPixelColor(imageData, x, y)
This is declared for you to use: const terrainColorBytes = [ rockColorBytes, dirtColorBytes ]

If you have other ideas on how to clean it up but keeping functionality, please do so. Also give
me any functions you needed to create for that refactoring.

The update function is:
...