Bugfix; height adjustment ChatGPT: default settings
>>> Prompt:
Given a game with a lemming sprite. It is traversing terrain on a 2D canvas. Gravity will always
pull it down, but bumpy terrain should keep it up. That is, it is not just walking in a straight
line along the X axis.

How should we adjust its position along the Y axis as it moves along the X axis?

You want to fix the code around heightAdjustment.

The lemming has the following code now:
...