Disable antialiasing, digger fixes, CSS fixes ChatGPT: default
>>> Prompt 1 CSS
Put this div in bottom left corner of the window: <div id="coordinatesDiv"></div>

>>> Prompt 2 Antialiasing off
give me a function that iterates over all pixels on a canvas and if a pixel does not
match one of the colors in terrainColorBytes, it should set the pixel to 0,0,0,255.
If the pixel is 0,0,0,255 already, you don't have do anything. This is already declared
globally:

const waterColorBytes = [0x00, 0x77, 0xbe]; // [0, 119, 190];
const rockColorBytes  = [0x88, 0x88, 0x88]; // [136, 136, 136];
const dirtColorBytes  = [0x4a, 0x2e, 0x00]; // [74, 46, 0];

const terrainColorBytes = [ rockColorBytes, dirtColorBytes ];

>>> Prompt 3 digger fixes
The function bash(lemming) below lets a lemming clear pixels in front of it along the X axis.
It does this by checking if a pixel is of a terrain color. I want a new function that does
exactly the same except that instead of checking and removing pixels in front of it along the
X axis, it should to it along the Y axis. In short: It should dig downwards (and never upwards,
so some code might be redundant in this case). Call this new function dig() instead of bash().
Give me the full function.

<< code for bash function here >>